\name{cartoPOO-package}
\alias{cartoPOO-package}
\alias{cartoPOO}
\docType{package}
\title{
A POO approach to cartography
}
\description{
This package makes some cartographic S4 classes and methods available.  
}
\details{
\tabular{ll}{
Package: \tab cartoPOO\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-05-25\cr
License: \tab GPL(>=2.0)\cr
}
bla bla}
\author{
Joël Gombin

Maintainer: Joël Gombin <joel.gombin@gmail.com>
}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[rCarto]{rCarto}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
