\docType{methods}
\name{inset}
\alias{inset}
\alias{show,inset-method}
\title{Create insets of some subparts of the map}
\arguments{
  \item{x}{a map.}

  \item{IDs}{List of vectors of the IDs of the polygons to
  be represented in each inset.}

  \item{coords}{Either a list of lists of the coordinates
  where the insets should be, or "locator" to interactively
  select the coordinates.}
}
\description{
  This method allows to create small maps (insets) of some
  subparts of the main map
}

