\docType{methods}
\name{plotHistLegend}
\alias{plotHistLegend}
\alias{plotHistLegend,map.choropleth-method}
\title{Plots a legend as a histogram}
\arguments{
  \item{carte}{an object of class map or one of its
  subclasses}
}
\description{
  This method plots a histogram-legend and is for internal
  use mainly.
}
\details{
  Currently, ploHistLegend only works with pdf or png
  devices, not with the X11 device.
}

